Changed naming convention for mainfunction queue and dlq becuase the default cdk names say autoalarmautoalarmautoalarm...
Changed mesageGroupId to 'autoalarm' for all eventbridge rules associated with the main function.

Removed unnecessary permissions for mainFunctionExecutionRole as we are granting consumption of the Fifo Queue messages to mainFunction in the creation of the fifo queue itself.

The main handler was broken and required a rewrite of the case switching in the handler function.