First implementation of logic to enable modular batch processing for EC2 events in preparation of the introduction of prometheus functionality. The New logic stores all instanceIDs and Tags into an array of objects which are looped through in manageActiveEC2 Alarms and manageInactiveEC2Alarm functions in ec2-modules.mts.
Additional changes were made to main-handler.mts to extract out the EC2 logic and how those events are handled. We are now consolidating all of those events into their own arrays which are then processed as whole objects for either tag changes or state change events.

New interface and type added to types.mts to create a EC2AlarmManagerArray which is used to type the array of instanceID and related tags in main-handler.mts ec2-modules.mts and later, prometheus-modules.

Untested.