Fixed some typos in prometheus-tools.mts
Adjusted the EC2AlarmmanagerObject interface to now include the state of the ec2 instance in preparation for prometheus in AutoAlarm which will help reduce duplicate logic in prometheus-tools.mts

Changes in both main-handler.mts and ec2-modules.mts to account for the changes needed/required for prometheus-tools.mts to pass all required information to delete or create prometheus alarms in batches.

Tested and working.