Working on implementing the snowball anti-pattern fix for reAlarm.
Added dedicated sqs queue and DLQ for the producer lambda so that the custom reAlarm rules can be processed in order along with scheduled triggers and provide better and more granular retry logic for customers and users who use this feature heavily.

Fixed a couple issues with unused vars/class attributes that are no longer used.

ReAlarm now uses FIFO queues with batching of up to 10 items in additona to reporting batch item failures in preparation for our bug fix to retry individual items vs the entire batch in the three handlers related to ReAlarm: realarm event rule handler, the producer handler and the consumer handler.

Updated the funciton definiton for realarm-event-rule-function.ts to use the sqs queue arn for the target used in the realarm-event-rule-handler.mts (which was updated as well to use the sqs queue arn)

untested and fix partially implemented. Pushing up for tracking.