Adjusted types to acoomodate prometheus confis without reinventing the wheel and utilizing what we currently have.
I only currently have been provided the metrics that are being exported for oracle DB and have created the config boilerplate foer those but still need the mysql and postress.

Todo at this point is create new alarm config functions to parse through tags that contain elements are isolated to prometheus and then correspond those to the correct metrics via an identifier which at this point appears to be 'engine'. scraping this from the secret seems straight forward. With this as well, I'll need to create a schema that creates a pattern for the host and time series that can be replaced by tag values. Is easy but requires uniformity across implementation.

Additionally, I'll need to create eventbridge rules that algin with the event types for secret manage. Also accounting for the fact that aws changes the casing for the key that connects to the ARN contained in those event triggers (easy to fix with toLower or whatever the JS equivalent is) as this is standard practice from AWS to protect secret manager.

Once all the above is done, I'll need to pull in the logic to set alarms from prometheus tools and confirm those work/are extensible to this use case.